Add files via upload
Adds Memory Checks and if using --lowvram or --medvram will process with CPU if you have low VRAM. 
Adds "threshold" which you can edit to set the videoram threshold for when you switch from video to CPU. 
Performs a basic memory check for torch. 
Note: not to be used when training models, because it is really slow!